+++ src/main/java/org/example/hakmana/LoginPageController.java
@@ -171,17 +171,22 @@ public class LoginPageController implements Initializable {
     }
 
     public void forgotPsswrdDialogPane(MouseEvent event) throws IOException{
+        ForgotPasswrdController forgotPasswrdController=new ForgotPasswrdController();
         FXMLLoader forgotFxmlLoad = new FXMLLoader();
         forgotFxmlLoad.setLocation(getClass().getResource("Scene/DialogBox/FrogotPasswrd.fxml"));
+        forgotFxmlLoad.setController(forgotPasswrdController);
         DialogPane forgotDialogPane=forgotFxmlLoad.load();
 
         Dialog<ButtonType> dialog=new Dialog<>();
         dialog.setDialogPane(forgotDialogPane);
         dialog.setTitle("Forgot password");
 
+
        Optional<ButtonType> clickedButton=dialog.showAndWait();
         if (clickedButton.isPresent() && clickedButton.get() == ButtonType.APPLY) {
             System.out.println("Apply");
+            forgotPasswrdController.newPsswrd();
+            clickedButton=dialog.showAndWait();
         }
     }
