+++ src/main/resources/org/example/hakmana/Scene/DialogBox/FrogotPasswrd.fxml
@@ -1,20 +1,22 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
 <?import javafx.geometry.Insets?>
+<?import javafx.scene.Cursor?>
 <?import javafx.scene.control.Button?>
 <?import javafx.scene.control.ButtonType?>
 <?import javafx.scene.control.DialogPane?>
 <?import javafx.scene.control.Label?>
 <?import javafx.scene.control.TextField?>
+<?import javafx.scene.effect.Glow?>
 <?import javafx.scene.image.Image?>
 <?import javafx.scene.image.ImageView?>
 <?import javafx.scene.layout.HBox?>
 <?import javafx.scene.layout.VBox?>
 <?import javafx.scene.text.Font?>
 
-<DialogPane minHeight="449.0" minWidth="546.0" prefHeight="502.0" prefWidth="546.0" style="-fx-background-color: CEEAF7;" xmlns="http://javafx.com/javafx/21" xmlns:fx="http://javafx.com/fxml/1" fx:controller="org.example.hakmana.ForgotPasswrdController">
+<DialogPane minHeight="449.0" minWidth="451.0" prefHeight="521.0" prefWidth="451.0" style="-fx-background-color: CEEAF7;" xmlns="http://javafx.com/javafx/21" xmlns:fx="http://javafx.com/fxml/1">
    <header>
-      <Label alignment="CENTER" text="Create new password" textAlignment="CENTER">
+      <Label alignment="CENTER" style="-fx-background-color: #D9D9D9; -fx-background-radius: 20;" text="Create new password" textAlignment="CENTER">
          <font>
             <Font name="System Bold" size="25.0" />
          </font>
@@ -28,16 +30,16 @@
       </Label>
    </header>
    <padding>
-      <Insets bottom="5.0" left="10.0" right="10.0" top="5.0" />
+      <Insets bottom="5.0" left="10.0" right="10.0" top="10.0" />
    </padding>
    <content>
-      <VBox prefHeight="200.0" prefWidth="100.0" spacing="20.0">
+      <VBox prefHeight="200.0" prefWidth="100.0" spacing="10.0">
          <children>
             <VBox fx:id="usrMailVbox" fillWidth="false" VBox.vgrow="NEVER">
                <children>
-                  <Label text="Enter your registered email" textFill="#59057b">
+                  <Label text="Enter your registered email" textFill="#0f0766">
                      <font>
-                        <Font size="18.0" />
+                        <Font size="16.0" />
                      </font>
                      <graphic>
                         <ImageView fitHeight="18.0" fitWidth="18.0" pickOnBounds="true" preserveRatio="true">
@@ -71,12 +73,15 @@
                      </font>
                   </Label>
                </children>
+               <VBox.margin>
+                  <Insets top="10.0" />
+               </VBox.margin>
             </VBox>
-            <VBox fx:id="verificationCodeVbox" disable="true" fillWidth="false" prefHeight="128.0" prefWidth="506.0" VBox.vgrow="NEVER">
+            <VBox fx:id="verificationCodeVbox" disable="true" fillWidth="false" prefHeight="128.0" prefWidth="506.0" style="-fx-border-width: 0 0 1 0; -fx-border-color: black;" VBox.vgrow="NEVER">
                <children>
-                  <Label text="Enter verification code" textFill="#59057b">
+                  <Label text="Enter verification code" textFill="#0f0766">
                      <font>
-                        <Font size="18.0" />
+                        <Font size="16.0" />
                      </font>
                      <graphic>
                         <ImageView fitHeight="18.0" fitWidth="18.0" pickOnBounds="true" preserveRatio="true">
@@ -106,16 +111,16 @@
                         <Insets />
                      </VBox.margin>
                   </HBox>
-                  <Button fx:id="verfiyBtn" mnemonicParsing="false" text="verify" />
+                  <Button fx:id="verfiyBtn" mnemonicParsing="false" onAction="#verifyOTP" text="verify" />
                </children>
             </VBox>
-            <VBox fx:id="newEmailVbox" disable="true" fillWidth="false" prefHeight="156.0" prefWidth="506.0" VBox.vgrow="NEVER">
+            <VBox fx:id="newPsswrdVbox" disable="true" fillWidth="false" prefHeight="156.0" prefWidth="506.0" VBox.vgrow="NEVER">
                <children>
                   <VBox spacing="8.0" VBox.vgrow="NEVER">
                      <children>
-                        <Label text="Enter new password" textFill="#59057b">
+                        <Label text="Enter new password" textFill="#0f0766">
                            <font>
-                              <Font size="18.0" />
+                              <Font name="System Italic" size="16.0" />
                            </font>
                            <graphic>
                               <ImageView fitHeight="18.0" fitWidth="18.0" pickOnBounds="true" preserveRatio="true">
@@ -134,9 +139,9 @@
                   </VBox>
                   <VBox spacing="8.0" VBox.vgrow="NEVER">
                      <children>
-                        <Label text="Re-enter new password" textFill="#59057b">
+                        <Label text="Re-enter new password" textFill="#0f0766">
                            <font>
-                              <Font size="18.0" />
+                              <Font name="System Italic" size="16.0" />
                            </font>
                            <graphic>
                               <ImageView fitHeight="18.0" fitWidth="18.0" pickOnBounds="true" preserveRatio="true">
@@ -162,4 +167,10 @@
       <ButtonType fx:constant="CANCEL" />
       <ButtonType fx:constant="APPLY" />
    </buttonTypes>
+   <effect>
+      <Glow level="0.19" />
+   </effect>
+   <cursor>
+      <Cursor fx:constant="HAND" />
+   </cursor>
 </DialogPane>
